Se usa Guice para injectar dependencia a la clase HandlersChannel
