+++ src/main/java/com/fox/platform/cntsrvex/infra/hndlr/HandlersChannel.java
@@ -1,0 +1,17 @@
+package com.fox.platform.cntsrvex.infra.hndlr;
+
+import io.vertx.ext.web.RoutingContext;
+
+/**
+ * Interface that defines the methods of the
+ * endpoint handlers implementation.
+ *
+ * @author juan.toro
+ *
+ */
+public interface HandlersChannel {
+
+
+  public void getChannelsAle(RoutingContext routingContext);
+  public void getChannelsJuan(RoutingContext routingContext);
+}
